requirements.txt
alembic==1.4.2
APScheduler==3.6.3
attrs==19.3.0
Babel==2.8.0
blis==0.4.1
catalogue==1.0.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorhash==1.0.2
configparser==5.0.0
cryptography==2.9.1
cymem==2.0.3
dateparser==0.7.4
decorator==4.4.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-MonitoringDashboard==3.0.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
future==0.18.2
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
Mako==1.1.2
MarkupSafe==1.1.1
murmurhash==1.0.2
numpy==1.18.3
plac==1.1.3
preshed==3.0.2
psutil==5.7.0
psycopg2==2.8.5
pycparser==2.20
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-telegram-bot==12.6.1
pytz==2019.3
PyYAML==5.3.1
regex==2020.4.4
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
spacy==2.2.4
SQLAlchemy==1.3.16
srsly==1.0.2
thinc==7.4.0
tornado==6.0.4
tqdm==4.45.0
tzlocal==2.0.0
urllib3==1.25.9
wasabi==0.6.0
Werkzeug==1.0.1
zipp==3.1.0
